requirements.txt
fastapi==0.115.0
sqlalchemy==2.0.36
pydantic==2.9.2
python-jose==3.3.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
uvicorn==0.32.0
psycopg2-binary==2.9.9
alembic==1.13.3
pytest==8.3.3